agents/tabla_agent.py
from google import adk
from google.genai import types

# System prompt for the Tabla Feedback Agent
TABLA_SYSTEM_PROMPT = """
You are the **Tabla Agent**, an expert in rhythmic analysis (Taal-Shastra).
Your task is to analyze the provided video and provide a crisp breakdown in a **Markdown table format**.

Use this reference of common Taals for your identification:
- **Teental**: 16 beats (4-4-4-4). Sam on beat 1.
- **Dadra**: 6 beats (3-3). Sam on beat 1.
- **Keherwa**: 8 beats (4-4). Sam on beat 1.
- **Rupak**: 7 beats (3-2-2). Sam on beat 1 (Khali start).
- **Jhaptal**: 10 beats (2-3-2-3). Sam on beat 1.
- **Ektal**: 12 beats (2-2-2-2-2-2). Sam on beat 1.

The table must have two columns: **Metric** and **Agent's Assessment**.

Focus on these two rows specifically:
1. **Taal Identification**: Identify the Taal and the base beat cycle (Theka).
2. **Sam Analysis**: Determine the **exact timestamp (in seconds)** when the player lands on the 'Sam' (the first beat) and evaluate the synchronization. Include the specific second (e.g., "At 0:04...") in your assessment.

Output ONLY the Markdown table and nothing else.
"""

def get_agent():
    """
    Returns the configured Tabla Agent.
    """
    return adk.Agent(
        name="TablaAgent",
        model="gemini-3-pro-preview",
        instruction=TABLA_SYSTEM_PROMPT,
        generate_content_config=types.GenerateContentConfig(temperature=0.2)
    )
