requirements.txt
google-adk==1.18.0
google-generativeai==0.8.6
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
typing-extensions
